# app/Http/Controllers/HomeController.php
<?php

declare(strict_types=1);

namespace App\Http\Controllers;

use App\Classes\Feed;
use App\Classes\SiteSearch;
use App\Classes\Validator;
use App\Models\Ban;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\RedirectResponse;
use Illuminate\Http\Request;
use Illuminate\View\View;
use Mobicms\Captcha\Image as MobicmsCaptcha;
use Symfony\Component\HttpFoundation\Response;
use Visavi\Captcha\CaptchaBuilder as AnimatedCaptchaBuilder;
use Visavi\Captcha\PhraseBuilder as AnimatedPhraseBuilder;

class HomeController extends Controller
{
    /**
     * Главная страница
     */
    public function index(): View
    {
        return view('index');
    }

    /**
     * Лента (AJAX)
     */
    public function feed(Request $request): string|RedirectResponse
    {
        if (! $request->ajax()) {
            return redirect(url('/') . ($request->has('page') ? '?page=' . $request->input('page') : ''));
        }

        return (string) (new Feed())->getFeed();
    }

    /**
     * Закрытие сайта
     */
    public function closed(): Response
    {
        if (setting('closedsite') !== 2) {
            return redirect('/');
        }

        return response()->view('pages/closed', [], 503);
    }

    /**
     * Поиск по сайту
     */
    public function search(Request $request, Validator $validator): View|RedirectResponse
    {
        $posts = paginate([], 10);
        $query = SiteSearch::clean((string) $request->input('query', $request->input('q', '')));
        $searchQuery = SiteSearch::terms($query);

        $types = SiteSearch::types();
        $sort = SiteSearch::sort(check($request->input('sort', 'relevance')));
        $type = SiteSearch::type(check($request->input('type')));

        if ($query) {
            $validator->length($searchQuery, SiteSearch::MIN_LENGTH, SiteSearch::MAX_LENGTH, ['find' => __('main.request_length')]);

            if ($validator->isValid()) {
                $posts = SiteSearch::paginate($searchQuery, $type, $sort, 10)
                    ->appends(compact('query', 'sort', 'type'));
            } else {
                // GET-страница рендерится сразу, редирект тут неуместен —
                // flash() пишет в сессию немедленно и корректно истекает
                session()->flash('danger', $validator->getErrors());
            }
        }

        return view('search/index', compact('posts', 'types', 'type', 'sort', 'query'));
    }

    /**
     * Бан по IP
     */
    public function ipban(Request $request): Response
    {
        $ban = Ban::query()
            ->where('ip', getIp())
            ->first();

        if (! $ban) {
            clearCache('ipBan');

            return redirect('/');
        }

        if (
            ! $ban->user_id
            && $ban->created_at->lt(now()->subMinute())
            && $request->isMethod('post')
            && captchaVerify()
        ) {
            $ban->delete();
            clearCache('ipBan');

            return redirect('/')
                ->with('success', __('pages.ip_success_unbanned'));
        }

        return response()->view('pages/ipban', compact('ban'), 429);
    }

    /**
     * Защитная картинка
     */
    public function captcha(Request $request): Response
    {
        if (setting('captcha_type') === 'animated') {
            $phrase = new AnimatedPhraseBuilder();
            $phrase = $phrase->getPhrase(setting('captcha_maxlength'), setting('captcha_symbols'));

            $captcha = new AnimatedCaptchaBuilder($phrase);
            $captcha = $captcha->render();
        } else {
            $captcha = new MobicmsCaptcha();
            $captcha->imageWidth = 180;
            $captcha->imageHeight = 50;
            $captcha->lengthMax = setting('captcha_maxlength');
            $captcha->characterSet = (string) setting('captcha_symbols');
            $phrase = $captcha->getCode();
            $captcha = $captcha->build();
        }

        $request->session()->put('protect', $phrase);

        return response($captcha)
            ->header('Content-Type', setting('captcha_type') === 'animated' ? 'image/gif' : 'image/png')
            ->header('Cache-Control', 'no-store, no-cache, must-revalidate, max-age=0')
            ->header('Pragma', 'no-cache')
            ->header('Expires', 'Sat, 26 Jul 1997 05:00:00 GMT');
    }

    /**
     * Быстрое изменение языка
     */
    public function language(string $lang, Request $request): JsonResponse
    {
        $languages = getAvailableLanguages();

        if (preg_match('/^[a-z]+$/', $lang) && in_array($lang, $languages, true)) {
            if ($user = $request->user()) {
                $user->update([
                    'language' => $lang,
                ]);
            } else {
                $request->session()->put('language', $lang);
            }
        }

        return response()->json(['success' => true]);
    }

    public function error403(): View
    {
        abort(403);
    }

    public function error404(): View
    {
        abort(404);
    }
}