# public/openapi/openapi.yaml
openapi: 3.2.0
info:
    title: 'Rotor API'
    version: 1.0.0
    description: |
        Для доступа к данным нужен API-ключ, который можно получить через авторизацию или на странице Мои данные.

        Лимиты запросов: 120 в минуту на все методы, 10 в минуту на авторизацию.
        Остаток лимита — в заголовках X-RateLimit-Limit и X-RateLimit-Remaining, при превышении приходит 429 и Retry-After.
security:
    -
        BearerToken: {  }
components:
    securitySchemes:
        BearerToken:
            type: http
            scheme: bearer
            description: 'Передайте API-ключ в заголовке: Authorization: Bearer {token}'
    parameters:
...
        Page:
            name: page
            in: query
            schema:
                type: integer
                default: 1
                minimum: 1
            description: 'Номер страницы'
...
    schemas:
...
        Author:
            type: object
            description: 'Автор записи в списках — подписи хватает аватара и ника'
            properties:
                login:
                    type: string
                name:
                    type:
                        - string
                        - 'null'
                level:
                    type: string
                color:
                    type:
                        - string
                        - 'null'
                    description: 'Цвет ника'
                avatar:
                    type:
                        - string
                        - 'null'
                    format: uri
                status:
                    type:
                        - string
                        - 'null'
                    description: 'HTML статуса пользователя'
...
        File:
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
                path:
                    type: string
                    format: uri
                size:
                    type: integer
                extension:
                    type: string
                mime_type:
                    type: string
                is_image:
                    type: boolean
                is_audio:
                    type: boolean
                is_video:
                    type: boolean
...
        FeedItem:
            type: object
            properties:
                type:
                    type: string
                    description: 'Тип записи (comments, news, topics, articles и т.д.)'
                id:
                    type: integer
                section:
                    type:
                        - string
                        - 'null'
                    description: 'Название раздела'
                title:
                    type:
                        - string
                        - 'null'
                url:
                    type:
                        - string
                        - 'null'
                    format: uri
                breadcrumbs:
                    type: array
                    description: 'Путь до записи от корня раздела, пустой массив — раздел не указан'
                    items:
                        type: object
                        properties:
                            title:
                                type: string
                            url:
                                type: string
                                format: uri
                text:
                    type:
                        - string
                        - 'null'
                    description: 'HTML текста записи'
                rating:
                    type:
                        - integer
                        - 'null'
                vote:
                    type:
                        - object
                        - 'null'
                    description: 'Цель голосования, null — тип не поддерживает рейтинг'
                    properties:
                        type:
                            type: string
                            description: 'Тип для POST /rating (у темы форума — последнее сообщение)'
                        id:
                            type: integer
                        value:
                            type:
                                - string
                                - 'null'
                            enum:
                                - +
                                - '-'
                                - null
                            description: 'Голос текущего пользователя, null без токена или если не голосовал'
                        own:
                            type: boolean
                            description: 'Своя запись — голосовать нельзя'
                comments_count:
                    type:
                        - integer
                        - 'null'
                user:
                    $ref: '#/components/schemas/Author'
                media:
                    type: array
                    description: 'Картинки и видео для галереи, кроме вставленных в текст'
                    items:
                        $ref: '#/components/schemas/File'
                files:
                    type: array
                    description: 'Остальные вложения — документы, архивы, аудио'
                    items:
                        $ref: '#/components/schemas/File'
                created_at:
                    type: string
                    format: date-time
                relate:
                    type: object
                    description: 'Только при type=comments: запись, к которой оставлен комментарий'
                    properties:
                        type:
                            type: string
                        id:
                            type: integer
                        title:
                            type:
                                - string
                                - 'null'
                        url:
                            type:
                                - string
                                - 'null'
                            format: uri
                price:
                    type:
                        - number
                        - 'null'
                    description: 'Только при type=items (объявления): цена, валюта — в настройке currency'
                phone:
                    type:
                        - string
                        - 'null'
                    description: 'Только при type=items (объявления): телефон продавца'
                offer_type:
                    type: string
                    enum:
                        - offer
                        - issue
                    description: 'Только при type=offers: предложение или проблема'
                status:
                    type: string
                    enum:
                        - wait
                        - process
                        - done
                        - cancel
                    description: 'Только при type=offers: статус обработки'
            additionalProperties: true
            description: |
                Набор полей зависит от type: базовые поля приходят всегда, дополнительные — только у своего типа записи
                (relate у комментариев, price и phone у объявлений, offer_type и status у предложений).
                Сторонний модуль может добавить собственные поля, поэтому неизвестные ключи игнорируйте, а не считайте ошибкой.
...
        SimplePaginationLinks:
            type: object
            description: 'last всегда null — общее число страниц не вычисляется'
            properties:
                first:
                    type: string
                    format: uri
                last:
                    type: 'null'
                prev:
                    type:
                        - string
                        - 'null'
                    format: uri
                next:
                    type:
                        - string
                        - 'null'
                    format: uri
        SimplePaginationMeta:
            type: object
            description: 'Упрощённая пагинация: общее число записей не считается, следующая страница определяется по links.next'
            properties:
                current_page:
                    type: integer
                current_page_url:
                    type: string
                    format: uri
                from:
                    type:
                        - integer
                        - 'null'
                path:
                    type: string
                    format: uri
                per_page:
                    type: integer
                to:
                    type:
                        - integer
                        - 'null'
...
paths:
...
    /feed:
        get:
            tags:
                - Лента
            summary: 'Лента последних записей сайта'
            description: 'Токен не обязателен. С токеном в каждом элементе приходит голос текущего пользователя (vote.value)'
            operationId: getFeed
            security:
                -
                    BearerToken: {  }
                - {}
            parameters:
                -
                    $ref: '#/components/parameters/Page'
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/FeedItem'
                                    links:
                                        $ref: '#/components/schemas/SimplePaginationLinks'
                                    meta:
                                        $ref: '#/components/schemas/SimplePaginationMeta'
                429:
                    description: 'Превышен общий лимит запросов'
...